Moves poincare visualization methods to new gensim.viz module
