+++ gensim/viz/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package contains functions to visualize different models from `gensim.models`.
+"""
